feat: 리뷰 등록 기능 구현 (#37)

+++ src/main/java/com/team01/deokhugam/review/dto/ReviewDto.java
@@ -11,9 +11,9 @@ public record ReviewDto(
     UUID userId,
     String userNickname,
     String content,
-    Double rating,
-    long likeCount,
-    long commentCount,
+    double rating,
+    int likeCount,
+    int commentCount,
     boolean likedByMe,
     OffsetDateTime createdAt,
     OffsetDateTime updatedAt
